test: disallow interrupts after cancel
This adjusts the tests to ensure that when `cancel` is
called in the replay engine, then this will make sure
that all subsequent interrupts are disregarded, too.

This new behaviour is necessary because otherwise we
would receive an interrupt but nothing was added to
the replay log (due to the cancel). This causes incorrect
state and will make the replay engine crash on subsequent
runs.